backend/requirements.txt
Django
djangorestframework
psycopg[binary]
langchain-google-genai
